Make GitLab tests pass
